feat: benchmark cache sizes and access latency
Iterate over a number of memory locations, measure the time elapsed and divide the time by the number of iterations. Namely, traverse a certain array using the random access pattern (from the hardware perspective) but in a deterministic way. In order to do so, we use a parameter called stride that represents the jump distance between memory locations.